+++ NEWS.md
@@ -2,6 +2,7 @@
 * Add com2links function that converts components to kinship links
 * Add tests for com2links
 * Add function to extract family tree from wiki family tree template
+* Add tests for readWikifamilytree
 
 # BGmisc 1.3.4.1
 * Hot fix to resolve issue with list of adjacency matrix not loading saved version